requirements/test.txt
# Test dependencies go here.
-r base.txt

coverage==4.4.1
django-coverage-plugin==1.5.0

# Pytest
pytest==3.0.7
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-sugar==0.8.0
